@@ -4,7 +4,6 @@ import com.typesafe.config.Config;
 import it.unimi.dsi.fastutil.longs.Long2ObjectOpenHashMap;
 import it.unimi.dsi.fastutil.shorts.Short2ObjectOpenHashMap;
 import org.gdal.ogr.*;
-import org.gdal.osr.CoordinateTransformation;
 import org.gdal.osr.SpatialReference;
 
 import java.io.File;
@@ -19,7 +18,7 @@ import java.util.stream.Stream;
 
 public class SingleCellConverter {
     private final boolean isValidCell;
-    private Logger logger = Logger.getLogger(CachedAdditionalDataSources.class.getName());
+    private final Logger logger = Logger.getLogger(SingleCellConverter.class.getName());
 
     private final File cellFile;
     private final Path outdir;
@@ -52,10 +51,11 @@ public class SingleCellConverter {
     private final Set<String> rightLetters = new HashSet<>(
             Arrays.asList("E", "F", "G", "H"));
 
-    private final Long2ObjectOpenHashMap<Node> nodes = new Long2ObjectOpenHashMap<>(50000);
-    private final Long2ObjectOpenHashMap<Way> ways = new Long2ObjectOpenHashMap<>(5000);
-    private final Long2ObjectOpenHashMap<Relation> relations = new Long2ObjectOpenHashMap<>(500);
-    private CoordinateTransformation srctowgs;
+    private final Long2ObjectOpenHashMap<LightNode> nodes = new Long2ObjectOpenHashMap<>(50000);
+    private final Long2ObjectOpenHashMap<LightWay> ways = new Long2ObjectOpenHashMap<>(5000);
+    private final Long2ObjectOpenHashMap<LightRelation> relations = new Long2ObjectOpenHashMap<>(500);
+    private GeomTransformer sphericToWGS;
+    private GeomTransformer srcToSphericMerc;
 
     SingleCellConverter(
             File cellFile,
@@ -137,19 +137,14 @@ public class SingleCellConverter {
     }
 
     private TagHandlerI getTagHandlerForDatasource(DataSource ds) {
-        switch (ds.GetLayer(0).GetName()) {
-            case "syvyyskayra_v":
-            case "syvyyspiste_p":
-                return syvyysTagHandler;
-            case "kesaretkeilyreitit":
-            case "ulkoilureitit":
-            case "luontopolut":
-            case "point_dump":
-                return retkeilyTagHandler;
-            default:
+        return switch (ds.GetLayer(0).GetName()) {
+            case "syvyyskayra_v", "syvyyspiste_p" -> syvyysTagHandler;
+            case "kesaretkeilyreitit", "ulkoilureitit", "luontopolut", "point_dump" -> retkeilyTagHandler;
+            default -> {
                 logger.severe("Unknown cached datasource ds name " + ds.GetLayer(0).GetName());
-                return null;
-        }
+                yield null;
+            }
+        };
     }
 
     private DataSource startReadingOGRFile(String fn) {
@@ -207,7 +202,7 @@ public class SingleCellConverter {
 
 
             ignored_fields.addAll(ignoredFields);
-            if (lyr.TestCapability(ogr.OLCIgnoreFields) && ignoredFields.size() > 0) {
+            if (lyr.TestCapability(ogr.OLCIgnoreFields) && !ignoredFields.isEmpty()) {
                 lyr.SetIgnoredFields(ignoredFields);
             }
 
@@ -223,7 +218,7 @@ public class SingleCellConverter {
                             logger.severe("NULL feature encountered on layer " + lyr.GetName());
                             return;
                         }
-                        if (!this.handleFeature(stringtable, lyr.GetName(), fieldMapping, feat, featurePreprocess, tagHandler)) {
+                        if (!this.handleFeature(stringtable, lyr.GetName(), fieldMapping, feat, tagHandler)) {
                             System.out.println("BREAK");
                             breakLayerLoop.set(true);
                         }
@@ -243,7 +238,7 @@ public class SingleCellConverter {
     }
 
     private boolean handleFeature(StringTable stringtable, String lyrname, ArrayList<Field> fieldMapping, Feature feat,
-                                  FeaturePreprocessI featurePreprocess, TagHandlerI tagHandler) {
+                                  TagHandlerI tagHandler) {
         Short2ObjectOpenHashMap<String> fields = new Short2ObjectOpenHashMap<>();
         Geometry geom;
         for (Field f : fieldMapping) {
@@ -256,13 +251,14 @@ public class SingleCellConverter {
 
         if (geom == null) return true;
 
-
-        geom = geom.SimplifyPreserveTopology(0.5);
-
-        if (srctowgs == null) {
+        if (sphericToWGS == null || srcToSphericMerc == null) {
             SpatialReference sref = geom.GetSpatialReference();
-            srctowgs = this.geomUtils.getTransformationToWGS84(sref.ExportToProj4());
+            sphericToWGS = this.geomUtils.spherictowgs;
+            srcToSphericMerc = this.geomUtils.getTransformationToSphereMercator(sref.ExportToProj4());
         }
+
+        geom = geom.Transform(srcToSphericMerc);
+        geom = geom.SimplifyPreserveTopology(0.5);
 
         GeomHandlerResult ghr;
 
@@ -285,9 +281,9 @@ public class SingleCellConverter {
 
         short tyyppi_value_id = stringtable.getStringId(tyyppi);
 
-        for (Node n : ghr.nodes) {
+        for (LightNode n : ghr.lightNodes) {
 
-            if (!n.isWaypart()) {
+            if (!n.isWayPart()) {
                 n.addTag(tyyppi_string_id, tyyppi_value_id);
                 tagHandler.addElementTags(n.nodeTags, fields, tyyppi, geomarea);
             }
@@ -297,7 +293,7 @@ public class SingleCellConverter {
             }
         }
 
-        for (Way w : ghr.ways) {
+        for (LightWay w : ghr.lightWays) {
             if (!w.getRole().equals("inner")) {
                 w.tags.put(tyyppi_string_id, tyyppi_value_id);
                 tagHandler.addElementTags(w.tags, fields, tyyppi, geomarea);
@@ -307,7 +303,7 @@ public class SingleCellConverter {
             }
         }
 
-        for (Relation r : ghr.relations) {
+        for (LightRelation r : ghr.lightRelations) {
 
             r.tags.put(tyyppi_string_id, tyyppi_value_id);
             tagHandler.addElementTags(r.tags, fields, tyyppi, geomarea);
@@ -323,61 +319,55 @@ public class SingleCellConverter {
 
         GeomHandlerResult ghr = new GeomHandlerResult();
 
-        if (geom.IsEmpty())
+        Geometry wgsgeom = geom.Transform(sphericToWGS);
+        if (geom.IsEmpty() || wgsgeom.IsEmpty()) {
             return ghr;
-
-        boolean ispoint = geom.GetGeometryType() == ogr.wkbPoint || geom.GetGeometryType() == ogr.wkbPoint25D;
-
-        double[][] srcpoints = geom.GetPoints();
-        double[][] wgspoints = geom.GetPoints();
-
-        srctowgs.TransformPoints(wgspoints);
-        Way w = null;
-        long wid;
-        if (!ispoint) {
-            wid = featureIDProvider.getWayID();
-            w = new Way();
-            w.id = wid;
         }
 
-        for (int i = 0; i < srcpoints.length; i++) {
+        boolean isPoint = geom.GetGeometryType() == ogr.wkbPoint || geom.GetGeometryType() == ogr.wkbPoint25D;
 
-            long phash = geomUtils.hashCoords(srcpoints[i][0], srcpoints[i][1]);
-            int pcell = geomUtils.xy2grid(srcpoints[i][0], srcpoints[i][1]);
+        LightWay w = null;
+        long wid;
+        if (!isPoint) {
+            wid = featureIDProvider.getWayID();
+            w = new LightWay(wid);
+        }
+
+        for (int i = 0; i < geom.GetPointCount(); i++) {
+
+            long phash = geomUtils.hashCoords(geom.GetX(i), geom.GetY(i));
+            int pcell = geomUtils.xy2grid(geom.GetX(i), geom.GetY(i));
 
             if (!nodes.containsKey(phash)) {
                 nodeCache.ensureGrid(pcell);
 
                 Optional<Long> cachedNodeId = nodeCache.getNodeId(pcell,phash);
 
-                long nodeid;
+                long nodeID;
                 if (cachedNodeId.isPresent()) {
-                    nodeid = cachedNodeId.get();
+                    nodeID = cachedNodeId.get();
                 } else {
-                    nodeid = featureIDProvider.getNodeID();
-                    if (this.nodeNearCellBorder(srcpoints[i])) {
-                        nodeCache.addNodeId(pcell, phash, nodeid);
+                    nodeID = featureIDProvider.getNodeID();
+                    if (this.nodeNearCellBorder(geom.GetPoint(i))) {
+                        nodeCache.addNodeId(pcell, phash, nodeID);
                     }
                 }
 
-                Node n = new Node(nodeid, phash, pcell, wgspoints[i][0], wgspoints[i][1], !ispoint);
+                LightNode n = new LightNode(nodeID, phash, wgsgeom.GetX(i), wgsgeom.GetY(i), !isPoint);
                 nodes.put(phash, n);
-                ghr.nodes.add(n);
-                if (!ispoint) {
-                    w.refs.add(n.getId());
-                }
+                ghr.lightNodes.add(n);
             } else {
-                Node n = nodes.get(phash);
-                n.waypart = n.waypart || !ispoint;
-                ghr.nodes.add(n);
-                if (!ispoint) {
-                    w.refs.add(n.getId());
-                }
+                LightNode n = nodes.get(phash);
+                n.wayPart = n.wayPart || !isPoint;
+                ghr.lightNodes.add(n);
+            }
+            if (!isPoint) {
+                w.refs.add(phash);
             }
         }
 
         if (w != null) {
-            ghr.ways.add(w);
+            ghr.lightWays.add(w);
         }
 
         return ghr;
@@ -405,14 +395,14 @@ public class SingleCellConverter {
             for (int i = 0; i < geom.GetGeometryCount(); i++) {
                 igeom = geom.GetGeometryRef(i);
                 ighr = this.handleSingleGeom(igeom);
-                ghr.nodes.addAll(ighr.nodes);
-                ghr.ways.addAll(ighr.ways);
+                ghr.lightNodes.addAll(ighr.lightNodes);
+                ghr.lightWays.addAll(ighr.lightWays);
             }
             return ghr;
         }
 
         long rid = featureIDProvider.getRelationID();
-        Relation r = new Relation();
+        LightRelation r = new LightRelation();
         r.setId(rid);
         r.tags.put(type, multipolygon);
 
@@ -420,23 +410,20 @@ public class SingleCellConverter {
             igeom = geom.GetGeometryRef(i);
 
             ighr = this.handleSingleGeom(igeom);
-            if (ighr.ways.size() == 0) {
+            if (ighr.lightWays.isEmpty()) {
                 return new GeomHandlerResult();
             }
-            ighr.ways.get(0).setRole((i == 0 ? "outer" : "inner"));
+            String role = i == 0 ? "outer" : "inner";
+            ighr.lightWays.getFirst().setRole(role);
 
-            ghr.nodes.addAll(ighr.nodes);
-            ghr.ways.addAll(ighr.ways);
+            ghr.lightNodes.addAll(ighr.lightNodes);
+            ghr.lightWays.addAll(ighr.lightWays);
 
-            RelationMember rm = new RelationMember();
-
-            rm.setId(ighr.ways.get(0).getId());
-            rm.setType();
-            rm.setRole((i == 0 ? "outer" : "inner"));
+            LightRelationMember rm = new LightRelationMember(ighr.lightWays.getFirst().getId(), role);
             r.members.add(rm);
         }
 
-        ghr.relations.add(r);
+        ghr.lightRelations.add(r);
         return ghr;
 
     }
