@@ -3,7 +3,7 @@ package org.hylly.mtk2garmin;
 import java.util.ArrayList;
 
 class GeomHandlerResult {
-    final ArrayList<Node> nodes = new ArrayList<>();
-    final ArrayList<Way> ways = new ArrayList<>();
-    final ArrayList<Relation> relations = new ArrayList<>();
+    final ArrayList<LightNode> lightNodes = new ArrayList<>();
+    final ArrayList<LightWay> lightWays = new ArrayList<>();
+    final ArrayList<LightRelation> lightRelations = new ArrayList<>();
 }
