@@ -3,7 +3,7 @@ package org.hylly.mtk2garmin;
 import it.unimi.dsi.fastutil.objects.ObjectOpenHashSet;
 import it.unimi.dsi.fastutil.shorts.Short2ObjectMap.Entry;
 import it.unimi.dsi.fastutil.shorts.Short2ObjectOpenHashMap;
-import it.unimi.dsi.fastutil.shorts.Short2ShortRBTreeMap;
+import it.unimi.dsi.fastutil.shorts.Short2ShortMap;
 
 import java.util.Objects;
 
@@ -56,7 +56,7 @@ class MMLTagHandler implements TagHandlerI {
     }
 
     @Override
-    public void addElementTags(Short2ShortRBTreeMap tags, Short2ObjectOpenHashMap<String> fields, String tyyppi, double geomarea) {
+    public void addElementTags(Short2ShortMap tags, Short2ObjectOpenHashMap<String> fields, String tyyppi, double geomarea) {
         if (tags.get(teksti_kieli) == fin && !Objects.equals(fields.get(teksti_kieli), "fin")) {
             return;
         }
@@ -66,7 +66,7 @@ class MMLTagHandler implements TagHandlerI {
             short kk = k.getShortKey();
             String val = k.getValue();
 
-            if (val.length() == 0) {
+            if (val.isEmpty()) {
                 continue;
             }
             if (kk == korarvo || kk == syvarvo) {
