@@ -3,7 +3,7 @@ package org.hylly.mtk2garmin;
 import it.unimi.dsi.fastutil.objects.ObjectOpenHashSet;
 import it.unimi.dsi.fastutil.shorts.Short2ObjectMap.Entry;
 import it.unimi.dsi.fastutil.shorts.Short2ObjectOpenHashMap;
-import it.unimi.dsi.fastutil.shorts.Short2ShortRBTreeMap;
+import it.unimi.dsi.fastutil.shorts.Short2ShortMap;
 
 import java.util.Arrays;
 
@@ -29,7 +29,7 @@ class ShapeRetkeilyTagHandler implements TagHandlerI {
     }
 
     @Override
-    public void addElementTags(Short2ShortRBTreeMap tags, Short2ObjectOpenHashMap<String> fields, String tyyppi, double geomarea) {
+    public void addElementTags(Short2ShortMap tags, Short2ObjectOpenHashMap<String> fields, String tyyppi, double geomarea) {
         for (Entry<String> k : fields.short2ObjectEntrySet()) {
             String val = k.getValue();
             
